ForecastVerification

ForecastVerification

ForecastVerification Julia package repo.

ForecastVerification.crms — Method.

crms(x,y)

Centred root mean square difference between x and y

source

ForecastVerification.rms — Method.

rms(x,y)

Root mean square difference between x and y

source

ForecastVerification.CRPS — Method.

crps = CRPS(X::AbstractMatrix,xobs::AbstractVector)

The Continuous Rank Probability Score (CRPS) from an ensemble X (cases,members) using the observation xobs (cases). The CRPS is averaged over all cases.

source

ForecastVerification.brierscore — Method.

BS = brierscore(y_true,y_prob::AbstractVector)

Compute the Brier Score for a series of binary events ytrue, forecasted with a probability `yprob`

$BS = \frac{1}{N}\sum_{t=1}^{N}(f_t - o_t)^2$

source

ForecastVerification.brierscoredecom — Method.

BS,REL,RES,UNC = brierscoredecom(y_true,y_prob::AbstractVector)

The Brier score can be decomposed into 3 additive components: Uncertainty, Reliability, and Resolution. (Murphy 1973)>

$BS = REL - RES + UNC$

$REL=\frac{1}{N} \sum_{k=1}^{K}{n_{k}(\mathbf{f_{k}} - \mathbf{\bar{o}}_{\mathbf{k}})}^{2}$

$RES = \frac{1}{N}\sum\limits _{k=1}^{K}{n_{k}(\mathbf{\bar{o}_{k}}-\bar{\mathbf{o}})}^{2}$

$UNC = \mathbf{\bar{o}}\left({1-\mathbf{\bar{o}}}\right)$

With $N$ being the total number of forecasts issued, $K$ the number of unique forecasts issued, $\mathbf{\bar{o}}={\sum_{t=1}^{N}}\mathbf{{o_t}}/N$ the observed climatological base rate for the event to occur, $n_{k}$ the number of forecasts with the same probability category and $\mathbf{\overline{o}}_{\mathbf{k}}$ the observed frequency, given forecasts of probability $\mathbf{f_{k}}$.

source